% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegKong-package.R
\docType{package}
\name{ggsegKong-package}
\alias{ggsegKong}
\alias{ggsegKong-package}
\title{ggsegKong: Kong Cortical Parcellation Atlases for the 'ggsegverse' Ecosystem}
\description{
Kong cortical parcellation atlases for the 'ggsegverse' ecosystem. Includes the Kong 2019 MS-HBM 17-network atlas and the Kong 2022 Areal MS-HBM 17-network parcellations at resolutions from 100 to 1000 parcels. Provides 'ggseg_atlas' objects with 2D polygon geometry and 3D vertex indices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegKong}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegKong/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
