% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kong2022.R
\name{kong2022_100}
\alias{kong2022_100}
\alias{kong2022_200}
\alias{kong2022_300}
\alias{kong2022_400}
\alias{kong2022_500}
\alias{kong2022_600}
\alias{kong2022_700}
\alias{kong2022_800}
\alias{kong2022_900}
\alias{kong2022_1000}
\title{Kong 2022 Areal MS-HBM 17-Network Atlases}
\usage{
kong2022_100()

kong2022_200()

kong2022_300()

kong2022_400()

kong2022_500()

kong2022_600()

kong2022_700()

kong2022_800()

kong2022_900()

kong2022_1000()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlases for the Kong et al. (2022) Areal MS-HBM cortical
parcellation aligned to Yeo 17 resting-state networks. Available in
100 to 1000 parcel configurations.
Contains both 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
kong2022_100()
plot(kong2022_100())
plot(kong2022_1000())
kong2022_200()
kong2022_300()
kong2022_400()
kong2022_500()
kong2022_600()
kong2022_700()
kong2022_800()
kong2022_900()
kong2022_1000()
}
\references{
Kong R, et al. (2022). "Individual-Specific Areal-Level
Parcellations Improve Functional Connectivity Prediction of Behavior."
\emph{Cerebral Cortex}, 31(10):4477-4500.
\doi{10.1093/cercor/bhab101}
}
\seealso{
Other ggseg_atlases: 
\code{\link{kong2019}()}

Other cortical_atlases: 
\code{\link{kong2019}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
